\name{NakaPT2CD-package}
\alias{NakaPT2CD-package}
\alias{NakaPT2CD}
\docType{package}
\title{
\packageTitle{NakaPT2CD}
}
\description{
\packageDescription{NakaPT2CD}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NakaPT2CD}
\packageIndices{NakaPT2CD}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{NakaPT2CD}

Maintainer: \packageMaintainer{NakaPT2CD}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
